requirement.txt
streamlit
scikit-learn
numpy
joblib
matplotlib
Pillow